Move generate-token into config module
