+++ primehub/__init__.py
@@ -209,7 +209,6 @@ class PrimeHub(object):
         self.register_command('files', 'Files')
         self.register_command('me', 'Me')
         self.register_command('version', 'Version')
-        self.register_command('tokens', 'TokenManager')
 
         # initial
         self._ensure_config_details(config)
